requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
gunicorn==23.0.0
psycopg[binary]==3.2.13
pgvector==0.3.6
redis==5.2.0
rq==1.16.2
requests==2.32.3
beautifulsoup4==4.12.3
trafilatura==1.12.2
lxml_html_clean==0.4.3
python-dotenv==1.0.1
openai==1.72.0
